// src/meta/common/interfaces.ts
export interface InstagramBasicDetailsParams {
    /** Whether to include followers count in the response. */
    followersCount?: boolean;
    /** Whether to include follows count in the response. */
    followsCount?: boolean;
    /** Whether to include biography in the response. */
    biography?: boolean;
    /** Whether to include media count in the response. */
    mediaCount?: boolean;
    /** Whether to include profile picture in the response. */
    profilePicture?: boolean;
    /** Whether to include website information in the response. */
    website?: boolean;
}


/**
 * Interface for parameters related to Instagram media data.
 */
export interface InstagramMediaParams {
    /** Limit the number of items returned. */
    limit?: number;
    /** Whether to include media URL in the response. */
    mediaUrl?: boolean;
    /** Whether to include caption in the response. */
    caption?: boolean;
    /** Whether to include comments count in the response. */
    commentsCount?: boolean;
    /** Whether to include ID information in the response. */
    id?: boolean;
    /** Whether to include Instagram ID in the response. */
    igId?: boolean;
    /** Whether to include comment enabled status in the response. */
    isCommentEnabled?: boolean;
    /** Whether to include shared to feed status in the response. */
    isSharedToFeed?: boolean;
    /** Whether to include like count in the response. */
    likeCount?: boolean;
    /** Whether to include media product type in the response. */
    mediaProductType?: boolean;
    /** Whether to include media type in the response. */
    mediaType?: boolean;
    /** Whether to include owner information in the response. */
    owner?: boolean;
    /** Whether to include permalink in the response. */
    permalink?: boolean;
    /** Whether to include shortcode in the response. */
    shortcode?: boolean;
    /** Whether to include thumbnail URL in the response. */
    thumbnailUrl?: boolean;
    /** Whether to include timestamp in the response. */
    timestamp?: boolean;
    /** Whether to include username in the response. */
    username?: boolean;
    /** Whether to include branded content partner promote information in the response. */
    brandedContentPartnerPromote?: boolean;
    /** Whether to include children in the response. */
    children?: boolean;
    /** Whether to include comments in the response. */
    comments?: boolean;
    /** Whether to include collaborators in the response. */
    collaborators?: boolean;
    /** Whether to include product tags in the response. */
    productTags?: boolean;
}